ActiScore Complete/AI Video Summarizer/requirements.txt
Flask
openai-whisper
pydub
reportlab
torch
ffmpeg-python